@@ -1,19 +1,23 @@
-# -*- coding: utf-8 -*-
-from unittest.mock import MagicMock, patch, ANY, call
+import json
+from unittest.mock import ANY, MagicMock, call, patch
 
 import pytest
-from kubernetes import stream
 from chaoslib.exceptions import ActivityFailed, InvalidActivity
 from chaoslib.provider.python import validate_python_activity
+from kubernetes import stream
 
-from chaosk8s.pod.actions import terminate_pods, exec_in_pods
-from chaosk8s.pod.probes import pods_in_phase, pods_not_in_phase, \
-    pods_in_conditions, all_pods_healthy
+from chaosk8s.pod.actions import exec_in_pods, terminate_pods
+from chaosk8s.pod.probes import (
+    all_pods_healthy,
+    pods_in_conditions,
+    pods_in_phase,
+    pods_not_in_phase,
+)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_pattern(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -30,17 +34,16 @@ def test_terminate_pods_by_name_pattern(cl, client, has_conf):
     client.CoreV1Api.return_value = v1
 
     ret = terminate_pods(name_pattern="my-app-[0-9]$")
-    
+
     assert len(ret) == 1
     assert ret == [pod.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod.metadata.name, "default", body=ANY)
+    v1.delete_namespaced_pod.assert_called_with(pod.metadata.name, "default", body=ANY)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_pattern_all(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -64,17 +67,19 @@ def test_terminate_pods_by_name_pattern_all(cl, client, has_conf):
     assert len(ret) == 2
     assert ret == [pod1.metadata.name, pod2.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
+    calls = [
+        call(pod1.metadata.name, "default", body=ANY),
+        call(pod2.metadata.name, "default", body=ANY),
+    ]
     v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_pattern_rand(cl, client, has_conf):
     # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
+    with patch("random.sample", side_effect=lambda seq, qty: seq[-qty:]):
         has_conf.return_value = False
         pod1 = MagicMock()
         pod1.metadata.name = "my-app-1"
@@ -101,12 +106,13 @@ def test_terminate_pods_by_name_pattern_rand(cl, client, has_conf):
         assert ret == [pod3.metadata.name]
         assert v1.delete_namespaced_pod.call_count == 1
         v1.delete_namespaced_pod.assert_called_with(
-            pod3.metadata.name, "default", body=ANY)
+            pod3.metadata.name, "default", body=ANY
+        )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_all(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -127,17 +133,19 @@ def test_terminate_pods_all(cl, client, has_conf):
     assert len(ret) == 2
     assert ret == [pod1.metadata.name, pod2.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
+    calls = [
+        call(pod1.metadata.name, "default", body=ANY),
+        call(pod2.metadata.name, "default", body=ANY),
+    ]
     v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_rand(cl, client, has_conf):
     # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
+    with patch("random.sample", side_effect=lambda seq, qty: seq[-qty:]):
         has_conf.return_value = False
         pod1 = MagicMock()
         pod1.metadata.name = "some-app"
@@ -158,12 +166,13 @@ def test_terminate_pods_rand(cl, client, has_conf):
         assert ret == [pod2.metadata.name]
         assert v1.delete_namespaced_pod.call_count == 1
         v1.delete_namespaced_pod.assert_called_with(
-            pod2.metadata.name, "default", body=ANY)
+            pod2.metadata.name, "default", body=ANY
+        )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_when_no_params_given(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -184,13 +193,12 @@ def test_terminate_pods_when_no_params_given(cl, client, has_conf):
     assert len(ret) == 1
     assert ret == [pod1.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod1.metadata.name, "default", body=ANY)
+    v1.delete_namespaced_pod.assert_called_with(pod1.metadata.name, "default", body=ANY)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_when_grace_period_is_set(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -209,13 +217,15 @@ def test_terminate_pods_when_grace_period_is_set(cl, client, has_conf):
     assert ret == [pod1.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 1
     v1.delete_namespaced_pod.assert_called_with(
-        pod1.metadata.name, "default",
-        body=client.V1DeleteOptions(grace_period_seconds=5))
+        pod1.metadata.name,
+        "default",
+        body=client.V1DeleteOptions(grace_period_seconds=5),
+    )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_given_percentage(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -237,22 +247,24 @@ def test_terminate_pods_by_given_percentage(cl, client, has_conf):
     v1.list_namespaced_pod.return_value = result
     client.CoreV1Api.return_value = v1
 
-    ret = terminate_pods(mode='percentage', qty=40)
+    ret = terminate_pods(mode="percentage", qty=40)
 
     assert len(ret) == 2
     assert ret == [pod1.metadata.name, pod2.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
+    calls = [
+        call(pod1.metadata.name, "default", body=ANY),
+        call(pod2.metadata.name, "default", body=ANY),
+    ]
     v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_given_percentage_rand(cl, client, has_conf):
     # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
+    with patch("random.sample", side_effect=lambda seq, qty: seq[-qty:]):
         has_conf.return_value = False
         pod1 = MagicMock()
         pod1.metadata.name = "some-app-1"
@@ -273,19 +285,21 @@ def test_terminate_pods_by_given_percentage_rand(cl, client, has_conf):
         v1.list_namespaced_pod.return_value = result
         client.CoreV1Api.return_value = v1
 
-        ret = terminate_pods(mode='percentage', qty=40, rand=True)
+        ret = terminate_pods(mode="percentage", qty=40, rand=True)
 
         assert len(ret) == 2
         assert ret == [pod3.metadata.name, pod4.metadata.name]
         assert v1.delete_namespaced_pod.call_count == 2
-        calls = [call(pod3.metadata.name, "default", body=ANY),
-                 call(pod4.metadata.name, "default", body=ANY)]
+        calls = [
+            call(pod3.metadata.name, "default", body=ANY),
+            call(pod4.metadata.name, "default", body=ANY),
+        ]
         v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_when_qty_grt_than_pods_selected(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -306,16 +320,19 @@ def test_terminate_pods_when_qty_grt_than_pods_selected(cl, client, has_conf):
     assert len(ret) == 2
     assert ret == [pod1.metadata.name, pod2.metadata.name]
     assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
+    calls = [
+        call(pod1.metadata.name, "default", body=ANY),
+        call(pod2.metadata.name, "default", body=ANY),
+    ]
     v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_should_fail_when_qty_less_than_pods_selected(
-        cl, client, has_conf):
+    cl, client, has_conf
+):
     has_conf.return_value = False
     pod1 = MagicMock()
     pod1.metadata.name = "some-app"
@@ -332,16 +349,14 @@ def test_terminate_pods_should_fail_when_qty_less_than_pods_selected(
 
     with pytest.raises(ActivityFailed) as excinfo:
         terminate_pods(qty=-4)
-    assert "Cannot select pods. Quantity '-4' is negative." in \
-        str(excinfo.value)
+    assert "Cannot select pods. Quantity '-4' is negative." in str(excinfo.value)
     assert v1.delete_namespaced_pod.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_should_fail_when_mode_not_present(
-        cl, client, has_conf):
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_terminate_pods_should_fail_when_mode_not_present(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
     pod1.metadata.name = "some-app"
@@ -358,14 +373,13 @@ def test_terminate_pods_should_fail_when_mode_not_present(
 
     with pytest.raises(ActivityFailed) as expinfo:
         terminate_pods(mode="some_mode")
-    assert "Cannot select pods. " \
-           "Mode 'some_mode' is invalid." in str(expinfo.value)
+    assert "Cannot select pods. " "Mode 'some_mode' is invalid." in str(expinfo.value)
     assert v1.delete_namespaced_pod.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
 def test_pods_in_phase(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -381,9 +395,9 @@ def test_pods_in_phase(cl, client, has_conf):
     assert pods_in_phase(label_selector="app=mysvc", phase="Running") is True
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
 def test_pods_should_have_been_phase(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -397,15 +411,15 @@ def test_pods_should_have_been_phase(cl, client, has_conf):
     client.CoreV1Api.return_value = v1
 
     with pytest.raises(ActivityFailed) as x:
-        assert pods_in_phase(
-            label_selector="app=mysvc", phase="Running") is True
-    assert "pod 'app=mysvc' is in phase 'Pending' but should be " \
-           "'Running'" in str(x.value)
+        assert pods_in_phase(label_selector="app=mysvc", phase="Running") is True
+    assert "pod 'app=mysvc' is in phase 'Pending' but should be " "'Running'" in str(
+        x.value
+    )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
 def test_pods_not_in_phase(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -418,13 +432,12 @@ def test_pods_not_in_phase(cl, client, has_conf):
     v1.list_namespaced_pod.return_value = result
     client.CoreV1Api.return_value = v1
 
-    assert pods_not_in_phase(
-        label_selector="app=mysvc", phase="Running") is True
+    assert pods_not_in_phase(label_selector="app=mysvc", phase="Running") is True
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
 def test_pods_in_conditions(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -440,38 +453,27 @@ def test_pods_in_conditions(cl, client, has_conf):
     client.CoreV1Api.return_value = v1
 
     # assert it works in the nominal case
-    assert pods_in_conditions(
-        label_selector="app=mysvc",
-        conditions=[
-            {
-                "type": "Ready",
-                "status": "True"
-            }
-        ]
-    ) is True
+    assert (
+        pods_in_conditions(
+            label_selector="app=mysvc", conditions=[{"type": "Ready", "status": "True"}]
+        )
+        is True
+    )
 
     # assert it works even if the hash is not in the same order
     # (just in case we add an ordered dict for some reason)
-    assert pods_in_conditions(
-        label_selector="app=mysvc",
-        conditions=[
-            {
-                "status": "True",
-                "type": "Ready"
-            }
-        ]
-    ) is True
+    assert (
+        pods_in_conditions(
+            label_selector="app=mysvc", conditions=[{"status": "True", "type": "Ready"}]
+        )
+        is True
+    )
 
     # assert it does not work when the condition is present but does not match
     with pytest.raises(ActivityFailed) as excinfo:
         pods_in_conditions(
             label_selector="app=mysvc",
-            conditions=[
-                {
-                    "status": "False",
-                    "type": "Ready"
-                }
-            ]
+            conditions=[{"status": "False", "type": "Ready"}],
         )
     assert str(excinfo)
 
@@ -479,477 +481,14 @@ def test_pods_in_conditions(cl, client, has_conf):
     with pytest.raises(ActivityFailed) as excinfo:
         pods_in_conditions(
             label_selector="app=mysvc",
-            conditions=[
-                {
-                    "status": "True",
-                    "type": "PodScheduled"
-                }
-            ]
+            conditions=[{"status": "True", "type": "PodScheduled"}],
         )
     assert str(excinfo)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_by_name_pattern(cl, client, has_conf):
-    has_conf.return_value = False
-    pod = MagicMock()
-    pod.metadata.name = "my-app-1"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(name_pattern="my-app-[0-9]$")
-
-    assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod.metadata.name, "default", body=ANY)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_by_name_pattern_all(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "my-app-1"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "my-app-2"
-
-    pod3 = MagicMock()
-    pod3.metadata.name = "my-app-test-3"
-
-    result = MagicMock()
-    result.items = [pod1, pod2, pod3]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(name_pattern="my-app-[0-9]$", all=True)
-
-    assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
-    v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_by_name_pattern_rand(cl, client, has_conf):
-    # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
-        has_conf.return_value = False
-        pod1 = MagicMock()
-        pod1.metadata.name = "my-app-1"
-
-        pod2 = MagicMock()
-        pod2.metadata.name = "my-app-2"
-
-        pod3 = MagicMock()
-        pod3.metadata.name = "my-app-3"
-
-        pod4 = MagicMock()
-        pod4.metadata.name = "my-app-test-1"
-
-        result = MagicMock()
-        result.items = [pod1, pod2, pod3, pod4]
-
-        v1 = MagicMock()
-        v1.list_namespaced_pod.return_value = result
-        client.CoreV1Api.return_value = v1
-
-        terminate_pods(name_pattern="my-app-[0-9]$", rand=True)
-
-        assert v1.delete_namespaced_pod.call_count == 1
-        v1.delete_namespaced_pod.assert_called_with(
-            pod3.metadata.name, "default", body=ANY)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_all(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod1, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(all=True)
-
-    assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
-    v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_rand(cl, client, has_conf):
-    # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
-        has_conf.return_value = False
-        pod1 = MagicMock()
-        pod1.metadata.name = "some-app"
-
-        pod2 = MagicMock()
-        pod2.metadata.name = "some-db"
-
-        result = MagicMock()
-        result.items = [pod1, pod2]
-
-        v1 = MagicMock()
-        v1.list_namespaced_pod.return_value = result
-        client.CoreV1Api.return_value = v1
-
-        terminate_pods(rand=True)
-
-        assert v1.delete_namespaced_pod.call_count == 1
-        v1.delete_namespaced_pod.assert_called_with(
-            pod2.metadata.name, "default", body=ANY)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_when_no_params_given(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod1, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods()
-
-    assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod1.metadata.name, "default", body=ANY)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_when_grace_period_is_set(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    result = MagicMock()
-    result.items = [pod1]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(grace_period=5)
-
-    assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod1.metadata.name, "default",
-        body=client.V1DeleteOptions(grace_period_seconds=5))
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_by_given_percentage(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app-1"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db-1"
-
-    pod3 = MagicMock()
-    pod3.metadata.name = "some-app-2"
-
-    pod4 = MagicMock()
-    pod4.metadata.name = "some-db-2"
-
-    result = MagicMock()
-    result.items = [pod1, pod2, pod3, pod4]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(mode='percentage', qty=40)
-
-    assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
-    v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_by_given_percentage_rand(cl, client, has_conf):
-    # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
-        has_conf.return_value = False
-        pod1 = MagicMock()
-        pod1.metadata.name = "some-app-1"
-
-        pod2 = MagicMock()
-        pod2.metadata.name = "some-db-1"
-
-        pod3 = MagicMock()
-        pod3.metadata.name = "some-app-2"
-
-        pod4 = MagicMock()
-        pod4.metadata.name = "some-db-2"
-
-        result = MagicMock()
-        result.items = [pod1, pod2, pod3, pod4]
-
-        v1 = MagicMock()
-        v1.list_namespaced_pod.return_value = result
-        client.CoreV1Api.return_value = v1
-
-        terminate_pods(mode='percentage', qty=40, rand=True)
-
-        assert v1.delete_namespaced_pod.call_count == 2
-        calls = [call(pod3.metadata.name, "default", body=ANY),
-                 call(pod4.metadata.name, "default", body=ANY)]
-        v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_when_qty_grt_than_pods_selected(cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod1, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    terminate_pods(qty=40)
-
-    assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
-    v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_should_fail_when_qty_less_than_pods_selected(
-        cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod1, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    with pytest.raises(ActivityFailed) as excinfo:
-        terminate_pods(qty=-4)
-    assert "Cannot select pods. Quantity '-4' is negative." in \
-        str(excinfo.value)
-    assert v1.delete_namespaced_pod.call_count == 0
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
-def test_terminate_pods_should_fail_when_mode_not_present(
-        cl, client, has_conf):
-    has_conf.return_value = False
-    pod1 = MagicMock()
-    pod1.metadata.name = "some-app"
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "some-db"
-
-    result = MagicMock()
-    result.items = [pod1, pod2]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    with pytest.raises(ActivityFailed) as expinfo:
-        terminate_pods(mode="some_mode")
-    assert "Cannot select pods. " \
-           "Mode 'some_mode' is invalid." in str(expinfo.value)
-    assert v1.delete_namespaced_pod.call_count == 0
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
-def test_pods_in_phase(cl, client, has_conf):
-    has_conf.return_value = False
-    pod = MagicMock()
-    pod.status = MagicMock()
-    pod.status.phase = "Running"
-    result = MagicMock()
-    result.items = [pod]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    assert pods_in_phase(label_selector="app=mysvc", phase="Running") is True
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
-def test_pods_should_have_been_phase(cl, client, has_conf):
-    has_conf.return_value = False
-    pod = MagicMock()
-    pod.status = MagicMock()
-    pod.status.phase = "Pending"
-    result = MagicMock()
-    result.items = [pod]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    with pytest.raises(ActivityFailed) as x:
-        assert pods_in_phase(
-            label_selector="app=mysvc", phase="Running") is True
-    assert "pod 'app=mysvc' is in phase 'Pending' but should be " \
-           "'Running'" in str(x.value)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
-def test_pods_not_in_phase(cl, client, has_conf):
-    has_conf.return_value = False
-    pod = MagicMock()
-    pod.status = MagicMock()
-    pod.status.phase = "Pending"
-    result = MagicMock()
-    result.items = [pod]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    assert pods_not_in_phase(
-        label_selector="app=mysvc", phase="Running") is True
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
-def test_pods_in_conditions(cl, client, has_conf):
-    has_conf.return_value = False
-    pod = MagicMock()
-    pod.status = MagicMock()
-    pod.status.conditions = [MagicMock()]
-    pod.status.conditions[0].type = "Ready"
-    pod.status.conditions[0].status = "True"
-    result = MagicMock()
-    result.items = [pod]
-
-    v1 = MagicMock()
-    v1.list_namespaced_pod.return_value = result
-    client.CoreV1Api.return_value = v1
-
-    # assert it works in the nominal case
-    assert pods_in_conditions(
-        label_selector="app=mysvc",
-        conditions=[
-            {
-                "type": "Ready",
-                "status": "True"
-            }
-        ]
-    ) is True
-
-    # assert it works even if the hash is not in the same order
-    # (just in case we add an ordered dict for some reason)
-    assert pods_in_conditions(
-        label_selector="app=mysvc",
-        conditions=[
-            {
-                "status": "True",
-                "type": "Ready"
-            }
-        ]
-    ) is True
-
-    # assert it does not work when the condition is present but does not match
-    with pytest.raises(ActivityFailed) as excinfo:
-        pods_in_conditions(
-            label_selector="app=mysvc",
-            conditions=[
-                {
-                    "status": "False",
-                    "type": "Ready"
-                }
-            ]
-        )
-    assert str(excinfo)
-
-    # assert it does not work when the condition is absent
-    with pytest.raises(ActivityFailed) as excinfo:
-        pods_in_conditions(
-            label_selector="app=mysvc",
-            conditions=[
-                {
-                    "status": "True",
-                    "type": "PodScheduled"
-                }
-            ]
-        )
-    assert str(excinfo)
-
-
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_by_name_pattern(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -981,17 +520,19 @@ def test_exec_in_pods_by_name_pattern(cl, client, has_conf):
     stream.stream.return_value.read_channel = MagicMock()
     stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
 
-    exec_in_pods(name_pattern="my-app-[0-9]$",
-                 cmd="dummy -a -b -c",
-                 container_name="container1",
-                 all=True)
+    exec_in_pods(
+        name_pattern="my-app-[0-9]$",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+        all=True,
+    )
 
     assert stream.stream.call_count == 2
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_by_name_pattern_rand(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -1027,18 +568,20 @@ def test_exec_in_pods_by_name_pattern_rand(cl, client, has_conf):
     stream.stream.return_value.read_channel = MagicMock()
     stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
 
-    exec_in_pods(name_pattern="my-app-[0-9]$",
-                 cmd="dummy -a -b -c",
-                 container_name="container2",
-                 qty=3,
-                 rand=True)
+    exec_in_pods(
+        name_pattern="my-app-[0-9]$",
+        cmd="dummy -a -b -c",
+        container_name="container2",
+        qty=3,
+        rand=True,
+    )
 
     assert stream.stream.call_count == 3
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
 def test_unhealthy_system_should_be_reported(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -1060,9 +603,9 @@ def test_unhealthy_system_should_be_reported(cl, client, has_conf):
     assert "the system is unhealthy" in str(excinfo.value)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_invalid_container_name(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -1087,36 +630,32 @@ def test_exec_in_pods_invalid_container_name(cl, client, has_conf):
     client.CoreV1Api.return_value = v1
 
     stream.stream = MagicMock()
-    exec_in_pods(name_pattern="my-app-[0-9]$",
-                 cmd="dummy -a -b -c",
-                 container_name="bad_container_name",
-                 qty=2,
-                 rand=True)
+    exec_in_pods(
+        name_pattern="my-app-[0-9]$",
+        cmd="dummy -a -b -c",
+        container_name="bad_container_name",
+        qty=2,
+        rand=True,
+    )
 
     assert stream.stream.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_no_command_provided(cl, client, has_conf):
     has_conf.return_value = False
 
     container1 = MagicMock()
     container1.name = "container1"
-    container2 = MagicMock()
-    container2.name = "container2"
 
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
-    pod1.spec.containers = [container1, container2]
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "my-app-2"
-    pod2.spec.containers = [container1, container2]
+    pod1.spec.containers = [container1]
 
     result = MagicMock()
-    result.items = [pod1, pod2]
+    result.items = [pod1]
 
     v1 = MagicMock()
     v1.list_namespaced_pod.return_value = result
@@ -1125,94 +664,86 @@ def test_exec_in_pods_no_command_provided(cl, client, has_conf):
     stream.stream = MagicMock()
 
     with pytest.raises(InvalidActivity) as expinfo:
-        validate_python_activity({
-            "name": "run-in-pod",
-            "type": "action",
-            "provider": {
-                "type": "python",
-                "module": "chaosk8s.pod.actions",
-                "func": "exec_in_pods",
-                "arguments": dict(
-                    name_pattern="my-app-[0-9]$",
-                    container_name="container1",
-                    qty=2,
-                    rand=True
-                )
+        validate_python_activity(
+            {
+                "name": "run-in-pod",
+                "type": "action",
+                "provider": {
+                    "type": "python",
+                    "module": "chaosk8s.pod.actions",
+                    "func": "exec_in_pods",
+                    "arguments": dict(
+                        name_pattern="my-app-[0-9]$",
+                        container_name="container1",
+                        qty=2,
+                        rand=True,
+                    ),
+                },
             }
-        })
+        )
 
     with pytest.raises(ActivityFailed) as expinfo:
         exec_in_pods(
             cmd=None,
             name_pattern="my-app-[0-9]$",
             container_name="container1",
-            qty=2,
-            rand=True
         )
 
     assert "A command must be set to run a container" in str(expinfo.value)
     assert stream.stream.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_no_pod_found_with_name(cl, client, has_conf):
     has_conf.return_value = False
 
     container1 = MagicMock()
     container1.name = "container1"
-    container2 = MagicMock()
-    container2.name = "container2"
 
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
-    pod1.spec.containers = [container1, container2]
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "my-app-2"
-    pod2.spec.containers = [container1, container2]
+    pod1.spec.containers = [container1]
 
     result = MagicMock()
-    result.items = [pod1, pod2]
+    result.items = [pod1]
 
     v1 = MagicMock()
     v1.list_namespaced_pod.return_value = result
     client.CoreV1Api.return_value = v1
 
     stream.stream = MagicMock()
-    exec_in_pods(name_pattern="no-app-[0-9]$",
-                 cmd="dummy -a -b -c",
-                 container_name="container1",
-                 qty=2,
-                 rand=True)
+    exec_in_pods(
+        name_pattern="no-app-[0-9]$",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+    )
 
     assert stream.stream.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_order_by_oldest(cl, client, has_conf):
     has_conf.return_value = False
 
     container1 = MagicMock()
     container1.name = "container1"
-    container2 = MagicMock()
-    container2.name = "container2"
 
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
-    pod1.metadata.creation_timestamp = "2019-11-25T13:50:31Z"
-    pod1.spec.containers = [container1, container2]
+    pod1.metadata.creation_timestamp = "2019-11-25T13:55:31Z"
+    pod1.spec.containers = [container1]
 
     pod2 = MagicMock()
     pod2.metadata.name = "my-app-2"
-    pod2.metadata.creation_timestamp = "2019-11-25T13:55:00Z"
-    pod2.spec.containers = [container1, container2]
+    pod2.metadata.creation_timestamp = "2019-11-25T13:50:00Z"
+    pod2.spec.containers = [container1]
 
     result = MagicMock()
-    result.items = [pod1, pod2]
+    result.items = [pod2, pod1]
 
     v1 = MagicMock()
     v1.list_namespaced_pod.return_value = result
@@ -1222,56 +753,51 @@ def test_exec_in_pods_order_by_oldest(cl, client, has_conf):
     stream.stream.return_value.read_channel = MagicMock()
     stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
 
-    exec_in_pods(name_pattern="my-app-[0-9]$",
-                 order="oldest",
-                 cmd="dummy -a -b -c",
-                 container_name="container1",
-                 qty=2,
-                 rand=True)
+    exec_in_pods(
+        name_pattern="my-app-[0-9]$",
+        order="oldest",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+        qty=2,
+    )
 
     assert stream.stream.call_count == 2
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_invalid_order(cl, client, has_conf):
     has_conf.return_value = False
 
     container1 = MagicMock()
     container1.name = "container1"
-    container2 = MagicMock()
-    container2.name = "container2"
 
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
-    pod1.spec.containers = [container1, container2]
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "my-app-2"
-    pod2.spec.containers = [container1, container2]
+    pod1.spec.containers = [container1]
 
     result = MagicMock()
-    result.items = [pod1, pod2]
+    result.items = [pod1]
 
     v1 = MagicMock()
     v1.list_namespaced_pod.return_value = result
     client.CoreV1Api.return_value = v1
     stream.stream = MagicMock()
     with pytest.raises(ActivityFailed) as excinfo:
-        exec_in_pods(name_pattern="my-app-[0-9]$",
-                     order="bad_order",
-                     cmd="dummy -a -b -c",
-                     container_name="container1",
-                     qty=2,
-                     rand=True)
-    assert "Cannot select pods. Order 'bad_order' is invalid." in  str(excinfo.value)
+        exec_in_pods(
+            name_pattern="my-app-[0-9]$",
+            order="bad_order",
+            cmd="dummy -a -b -c",
+            container_name="container1",
+        )
+    assert "Cannot select pods. Order 'bad_order' is invalid." in str(excinfo.value)
     assert stream.stream.call_count == 0
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_using_pod_label_selector(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -1280,7 +806,9 @@ def test_exec_in_pods_using_pod_label_selector(cl, client, has_conf):
 
     container2 = MagicMock()
     container2.name = "container2"
-    container2.metadata.labels = {"dummy_label1": "dummy_value1", "dummy_label1": "dummy_value2"}
+    container2.metadata.labels = {
+        "dummy_label1": "dummy_value1",
+    }
 
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
@@ -1301,34 +829,27 @@ def test_exec_in_pods_using_pod_label_selector(cl, client, has_conf):
     stream.stream.return_value.read_channel = MagicMock()
     stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
 
-    exec_in_pods(label_selector="dummy_label1=dummy_value1",
-                 cmd="dummy -a -b -c",
-                 container_name="container1",
-                 qty=2,
-                 rand=True)
+    exec_in_pods(
+        label_selector="dummy_label1=dummy_value1",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+        qty=2,
+    )
     assert stream.stream.call_count == 1
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_exec_in_pods_return_value(cl, client, has_conf):
     has_conf.return_value = False
 
     container1 = MagicMock()
     container1.name = "container1"
 
-    container2 = MagicMock()
-    container2.name = "container2"
-    container2.metadata.labels = {"dummy_label1": "dummy_value1", "dummy_label1": "dummy_value2"}
-
     pod1 = MagicMock()
     pod1.metadata.name = "my-app-1"
-    pod1.spec.containers = [container1, container2]
-
-    pod2 = MagicMock()
-    pod2.metadata.name = "my-app-2"
-    pod2.spec.containers = [container1, container2]
+    pod1.spec.containers = [container1]
 
     result = MagicMock()
     result.items = [pod1]
@@ -1341,19 +862,90 @@ def test_exec_in_pods_return_value(cl, client, has_conf):
     stream.stream.return_value.read_channel = MagicMock()
     stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
 
-    exec_in_pods(label_selector="dummy_label1=dummy_value1",
-                 cmd="dummy -a -b -c",
-                 container_name="container1",
-                 qty=2,
-                 rand=True)
+    exec_in_pods(
+        name_pattern="my-app-[0-9]$",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+    )
     assert stream.stream.call_count == 1
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.probes.client', autospec=True)
-@patch('chaosk8s.client')
-def test_succeeded_and_running_pods_should_be_considered_healthy(cl, client,
-                                                                 has_conf):
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_exec_in_pods_cmd_as_string(cl, client, has_conf):
+    has_conf.return_value = False
+
+    container1 = MagicMock()
+    container1.name = "container1"
+
+    pod1 = MagicMock()
+    pod1.metadata.name = "my-app-1"
+    pod1.spec.containers = [container1]
+
+    result = MagicMock()
+    result.items = [pod1]
+
+    v1 = MagicMock()
+    v1.list_namespaced_pod.return_value = result
+    client.CoreV1Api.return_value = v1
+
+    stream.stream = MagicMock()
+    stream.stream.return_value.read_channel = MagicMock()
+    stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
+
+    exec_in_pods(
+        name_pattern="my-app-1",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+    )
+    assert stream.stream.call_count == 1
+    assert stream.stream.call_args[1]["command"] == ["dummy", "-a", "-b", "-c"]
+
+
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_exec_in_pods_cmd_as_list(cl, client, has_conf):
+    has_conf.return_value = False
+
+    container1 = MagicMock()
+    container1.name = "container1"
+
+    pod1 = MagicMock()
+    pod1.metadata.name = "my-app-1"
+    pod1.spec.containers = [container1]
+
+    result = MagicMock()
+    result.items = [pod1]
+
+    v1 = MagicMock()
+    v1.list_namespaced_pod.return_value = result
+    client.CoreV1Api.return_value = v1
+
+    stream.stream = MagicMock()
+    stream.stream.return_value.read_channel = MagicMock()
+    stream.stream.return_value.read_channel.side_effect = [
+        "hello",
+        '{"status":"Success"}',
+    ]
+
+    command = ["dummy", "-l", "Long and 'complex'"]
+
+    result = exec_in_pods(
+        name_pattern="my-app-1",
+        cmd=command,
+        container_name="container1",
+    )
+    assert stream.stream.call_count == 1
+    assert stream.stream.call_args[1]["command"] == command
+    assert result[0]["exit_code"] == 0
+
+
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.probes.client", autospec=True)
+@patch("chaosk8s.client")
+def test_succeeded_and_running_pods_should_be_considered_healthy(cl, client, has_conf):
     has_conf.return_value = False
 
     podSucceeded = MagicMock()
@@ -1373,9 +965,9 @@ def test_succeeded_and_running_pods_should_be_considered_healthy(cl, client,
     assert health
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_with_prefix_pattern(cl, client, has_conf):
     has_conf.return_value = False
     pod = MagicMock()
@@ -1394,13 +986,12 @@ def test_terminate_pods_by_name_with_prefix_pattern(cl, client, has_conf):
     terminate_pods(name_pattern="my-app-[0-9]$")
 
     assert v1.delete_namespaced_pod.call_count == 1
-    v1.delete_namespaced_pod.assert_called_with(
-        pod.metadata.name, "default", body=ANY)
+    v1.delete_namespaced_pod.assert_called_with(pod.metadata.name, "default", body=ANY)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_with_prefix_pattern_all(cl, client, has_conf):
     has_conf.return_value = False
     pod1 = MagicMock()
@@ -1422,17 +1013,19 @@ def test_terminate_pods_by_name_with_prefix_pattern_all(cl, client, has_conf):
     terminate_pods(name_pattern="my-app-[0-9]$", all=True)
 
     assert v1.delete_namespaced_pod.call_count == 2
-    calls = [call(pod1.metadata.name, "default", body=ANY),
-             call(pod2.metadata.name, "default", body=ANY)]
+    calls = [
+        call(pod1.metadata.name, "default", body=ANY),
+        call(pod2.metadata.name, "default", body=ANY),
+    ]
     v1.delete_namespaced_pod.assert_has_calls(calls, any_order=True)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.pod.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_terminate_pods_by_name_with_prefix_pattern_rand(cl, client, has_conf):
     # Patch `random.sample` to always return last items in seq
-    with patch('random.sample', side_effect=lambda seq, qty: seq[-qty:]):
+    with patch("random.sample", side_effect=lambda seq, qty: seq[-qty:]):
         has_conf.return_value = False
         pod1 = MagicMock()
         pod1.metadata.name = "some-random-1-string-my-app-1"
@@ -1457,5 +1050,117 @@ def test_terminate_pods_by_name_with_prefix_pattern_rand(cl, client, has_conf):
 
         assert v1.delete_namespaced_pod.call_count == 1
         v1.delete_namespaced_pod.assert_called_with(
-            pod3.metadata.name, "default", body=ANY)
+            pod3.metadata.name, "default", body=ANY
+        )
 
+
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_exec_in_pods_stderr_can_be_json(cl, client, has_conf):
+    has_conf.return_value = False
+
+    container1 = MagicMock()
+    container1.name = "container1"
+
+    pod1 = MagicMock()
+    pod1.metadata.name = "my-app-1"
+    pod1.spec.containers = [container1]
+
+    result = MagicMock()
+    result.items = [pod1]
+
+    v1 = MagicMock()
+    v1.list_namespaced_pod.return_value = result
+    client.CoreV1Api.return_value = v1
+
+    stream.stream = MagicMock()
+    stream.stream.return_value.read_channel = MagicMock()
+    stream.stream.return_value.read_channel.return_value = json.dumps(
+        {
+            "status": "Error",
+            "message": "failed",
+            "details": {"causes": [{"message": 2}]},
+        }
+    )
+
+    result = exec_in_pods(
+        name_pattern="my-app-1",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+    )
+    assert stream.stream.call_count == 1
+    assert stream.stream.call_args[1]["command"] == ["dummy", "-a", "-b", "-c"]
+    assert result[0]["exit_code"] == 2
+    assert result[0]["stderr"] == "failed"
+
+
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_exec_in_pods_stderr_sometimes_not_json(cl, client, has_conf):
+    has_conf.return_value = False
+
+    container1 = MagicMock()
+    container1.name = "container1"
+
+    pod1 = MagicMock()
+    pod1.metadata.name = "my-app-1"
+    pod1.spec.containers = [container1]
+
+    result = MagicMock()
+    result.items = [pod1]
+
+    v1 = MagicMock()
+    v1.list_namespaced_pod.return_value = result
+    client.CoreV1Api.return_value = v1
+
+    stream.stream = MagicMock()
+    stream.stream.return_value.read_channel = MagicMock()
+    stream.stream.return_value.read_channel.return_value = "BOOM"
+
+    result = exec_in_pods(
+        name_pattern="my-app-1",
+        cmd="dummy -a -b -c",
+        container_name="container1",
+    )
+    assert stream.stream.call_count == 1
+    assert stream.stream.call_args[1]["command"] == ["dummy", "-a", "-b", "-c"]
+    assert result[0]["exit_code"] == 1
+    assert result[0]["stderr"] == "BOOM"
+
+
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.pod.actions.client", autospec=True)
+@patch("chaosk8s.client")
+def test_exec_in_pods_cmd_as_shlex(cl, client, has_conf):
+    has_conf.return_value = False
+
+    container1 = MagicMock()
+    container1.name = "container1"
+
+    pod1 = MagicMock()
+    pod1.metadata.name = "my-app-1"
+    pod1.spec.containers = [container1]
+
+    result = MagicMock()
+    result.items = [pod1]
+
+    v1 = MagicMock()
+    v1.list_namespaced_pod.return_value = result
+    client.CoreV1Api.return_value = v1
+
+    stream.stream = MagicMock()
+    stream.stream.return_value.read_channel = MagicMock()
+    stream.stream.return_value.read_channel.return_value = '{"status":"Success"}'
+
+    command = "/bin/sh -c ls"
+
+    result = exec_in_pods(
+        name_pattern="my-app-1",
+        cmd=command,
+        container_name="container1",
+    )
+    assert stream.stream.call_count == 1
+    assert stream.stream.call_args[1]["command"] == ["/bin/sh", "-c", "ls"]
+    assert result[0]["exit_code"] == 0
